requirements.txt
discord.py==2.6.4
aiosqlite==0.22.1
pyyaml==6.0.3